♻️ refactor: dropdown으로 분리했던 KeywordRanking 컴포넌트 합치기
